Various changes to a) integrate correctly with API GW, b) correctly clean-up DB sessions
