+++ config.ini
@@ -36,7 +36,7 @@ db_driver = postgresql
 db_user = postgres
 db_pass = postgres
 db_port = 5432
-db_name = onsdb
+db_name = postgres
 db_host = localhost
 ;
 ;   api_gateway needs to point at the API gateway you're using. Typically we
@@ -76,6 +76,9 @@ jwt_secret = vrwgLNWEffe45thh545yuby
 ;   5. int  - is the CF int environment
 ;   6. prod - is the CF production environment
 ;
+db_connection = ${db_driver}://${db_user}:${db_pass}@${db_host}:${db_port}/${db_name}
+db_schema = ras
+
 [test]
 drop_database = yes
 db_connection = sqlite:///${db_name}
