+++ swagger_server/controllers_local/controller.py
@@ -6,12 +6,16 @@ import uuid
 from flask import make_response, jsonify
 from ons_ras_common import ons_env
 
+from swagger_server.controllers_local.session_context import transaction
+from swagger_server.controllers_local.error_decorator import convert_exceptions
 from swagger_server.controllers_local.validate import Validator, Exists, IsUuid, IsIn
-from swagger_server.models._models import Business, Party, Respondent, BusinessRespondent
+from swagger_server.models.models import Business, Party, Respondent, BusinessRespondent
+
 
 db = ons_env.db
 
 
+@convert_exceptions
 def businesses_post(business):
     """
     adds a reporting unit of type Business
@@ -22,66 +26,62 @@ def businesses_post(business):
     :rtype: None
     """
 
-    v = Validator(Exists('businessRef',
-                         'attributes',
-                         'attributes.contactName',
-                         'attributes.employeeCount',
-                         'attributes.enterpriseName',
-                         'attributes.enterpriseName',
-                         'attributes.facsimile',
-                         'attributes.fulltimeCount',
-                         'attributes.legalStatus',
-                         'attributes.name',
-                         'attributes.sic2003',
-                         'attributes.sic2007',
-                         'attributes.telephone',
-                         'attributes.tradingName',
-                         'attributes.turnover'
-                         ))
-    if 'id' in business:
-        v.add_rule(IsUuid('id'))
-    if not v.validate(business):
-        return make_response(jsonify(v.errors), 400)
+    with transaction() as tran:
+        v = Validator(Exists('businessRef',
+                             'attributes',
+                             'attributes.contactName',
+                             'attributes.employeeCount',
+                             'attributes.enterpriseName',
+                             'attributes.enterpriseName',
+                             'attributes.facsimile',
+                             'attributes.fulltimeCount',
+                             'attributes.legalStatus',
+                             'attributes.name',
+                             'attributes.sic2003',
+                             'attributes.sic2007',
+                             'attributes.telephone',
+                             'attributes.tradingName',
+                             'attributes.turnover'
+                             ))
+        if 'id' in business:
+            v.add_rule(IsUuid('id'))
+        if not v.validate(business):
+            return make_response(jsonify(v.errors), 400)
 
-    party_uuid = business.get('id', uuid.uuid4())
-    business_ref = business['businessRef']
-    associations = business.get('associations')
+        party_uuid = business.get('id', uuid.uuid4())
+        business_ref = business['businessRef']
+        associations = business.get('associations')
 
-    db_party = db.session.query(Party).filter(Party.party_uuid == party_uuid).first()
+        db_party = tran.query(Party).filter(Party.party_uuid == party_uuid).first()
 
-    if db_party and not db_party.business:
-        return make_response(jsonify({'errors': "Existing party with '{}' does not identify a business."
-                                     .format(party_uuid)}), 400)
+        if db_party and not db_party.business:
+            return make_response(jsonify({'errors': "Existing party with '{}' does not identify a business."
+                                         .format(party_uuid)}), 400)
 
-    if db_party:
-        db_business = db_party.business
-    else:
-        db_party = Party(party_uuid)
-        db_business = Business(business_ref, db_party)
-        db.session.add(db_business)
+        if db_party:
+            db_business = db_party.business
+        else:
+            db_party = Party(party_uuid)
+            db_business = Business(business_ref, db_party)
+            tran.add(db_business)
 
-    db_business.from_dict(business['attributes'])
+        db_business.from_dict(business['attributes'])
 
-    if associations:
-        for assoc in associations:
-            assoc_id = assoc['id']
+        if associations:
+            for assoc in associations:
+                assoc_id = assoc['id']
 
-            assoc_party = db.session.query(Party).filter(Party.party_uuid == assoc_id).first()
-            business_respondent = BusinessRespondent()
-            business_respondent.respondent = assoc_party.respondent
-            business_respondent.business = db_business
-            # TODO: check the association doesn't already exist?
-            db.session.add(business_respondent)
+                assoc_party = tran.query(Party).filter(Party.party_uuid == assoc_id).first()
+                business_respondent = BusinessRespondent()
+                business_respondent.respondent = assoc_party.respondent
+                business_respondent.business = db_business
+                # TODO: check the association doesn't already exist?
+                tran.add(business_respondent)
 
-    # TODO: handle DB errors
-    db.session.commit()
-
-    return make_response(jsonify(db_business.to_dict()), 200)
+        return make_response(jsonify(db_business.to_dict()), 200)
 
 
-#
-# /businesses/ref/{ref}
-#
+@convert_exceptions
 def get_business_by_ref(ref):
     """
     Get a Business by its unique business reference
@@ -99,9 +99,7 @@ def get_business_by_ref(ref):
     return make_response(jsonify(business.to_dict()), 200)
 
 
-#
-# /businesses/id/{id}
-#
+@convert_exceptions
 def get_business_by_id(id):
     """
     Get a Business by its Party ID
@@ -125,9 +123,7 @@ def get_business_by_id(id):
     return make_response(jsonify(party.business.to_dict()), 200)
 
 
-#
-# /parties
-#
+@convert_exceptions
 def parties_post(party):
     """
     given a sampleUnitType B | H this adds a reporting unit of type Business or Household
@@ -148,9 +144,7 @@ def parties_post(party):
         return respondents_post(party)
 
 
-#
-# /parties/ref/{ref}
-#
+@convert_exceptions
 def get_party_by_ref(sampleUnitType, sampleUnitRef):
     """
     Get a Party by its unique reference (ruref / uprn)
@@ -168,9 +162,7 @@ def get_party_by_ref(sampleUnitType, sampleUnitRef):
         return get_business_by_ref(sampleUnitRef)
 
 
-#
-# /parties/{id}
-#
+@convert_exceptions
 def get_party_by_id(sampleUnitType, id):
     v = Validator(IsIn('sampleUnitType', 'B', 'BI'))
     if not v.validate({'sampleUnitType': sampleUnitType}):
@@ -182,9 +174,7 @@ def get_party_by_id(sampleUnitType, id):
         return get_respondent_by_id(id)
 
 
-#
-# /respondents/id/{id}
-#
+@convert_exceptions
 def get_respondent_by_id(id):
     """
     Get a Respondent by its Party ID
@@ -208,9 +198,7 @@ def get_respondent_by_id(id):
     return make_response(jsonify(party.respondent.to_dict()), 200)
 
 
-#
-# /respondents
-#
+@convert_exceptions
 def respondents_post(party):
     """
     adds a Respondent
@@ -220,31 +208,31 @@ def respondents_post(party):
 
     :rtype: None
     """
-    v = Validator(Exists('emailAddress', 'firstName', 'lastName', 'telephone'))
-    if 'id' in party:
-        v.add_rule(IsUuid('id'))
-    if not v.validate(party):
-        return make_response(jsonify(v.errors), 400)
 
-    party_uuid = party.get('id', uuid.uuid4())
-    db_party = db.session.query(Party).filter(Party.party_uuid == party_uuid).first()
+    with transaction() as tran:
+        v = Validator(Exists('emailAddress', 'firstName', 'lastName', 'telephone'))
+        if 'id' in party:
+            v.add_rule(IsUuid('id'))
+        if not v.validate(party):
+            return make_response(jsonify(v.errors), 400)
 
-    if db_party and not db_party.respondent:
-        return make_response(jsonify({'errors': "Existing party with '{}' does not identify a respondent."
-                                     .format(party_uuid)}), 400)
+        party_uuid = party.get('id', uuid.uuid4())
+        db_party = tran.query(Party).filter(Party.party_uuid == party_uuid).first()
 
-    if not db_party:
-        db_party = Party(party_uuid)
-        respondent = Respondent(db_party)
-        db.session.add(respondent)
-    else:
-        respondent = db_party.respondent
+        if db_party and not db_party.respondent:
+            return make_response(jsonify({'errors': "Existing party with '{}' does not identify a respondent."
+                                         .format(party_uuid)}), 400)
 
-    respondent.email_address = party['emailAddress']
-    respondent.first_name = party['firstName']
-    respondent.last_name = party['lastName']
-    respondent.telephone = party['telephone']
+        if not db_party:
+            db_party = Party(party_uuid)
+            respondent = Respondent(db_party)
+            tran.add(respondent)
+        else:
+            respondent = db_party.respondent
 
-    db.session.commit()
+        respondent.email_address = party['emailAddress']
+        respondent.first_name = party['firstName']
+        respondent.last_name = party['lastName']
+        respondent.telephone = party['telephone']
 
-    return make_response(jsonify(respondent.to_dict()), 200)
+        return make_response(jsonify(respondent.to_dict()), 200)
