+++ swagger_server/swagger/swagger.yaml
@@ -76,7 +76,7 @@ definitions:
       href: {example: /party/id/12345678, type: string}
       rel: {example: self, type: string}
       type: {example: application/vnd.ons.party+json, type: string}
-host: localhost:8000
+host: api-int.apps.mvp.onsclofo.uk:443
 info:
   contact: {email: ras@ons.gov.uk}
   description: Initial API for the Party microservice.
